yes, I am breaking everything; no, I am not sorry
a long time ago, on a hard disk far far away...
I wrote a network based DCPU emulator in C,
and so it was for many a year.
Until one day I got sick of looking at the crazy data model,
and all the extra lines to make it work, and thus it was so.
so much and overwhelming, that I tore it to bits, and now it's C++11 ish.
This lonely road being walked is now a little brighter, more object
oriented, and closer to being a program worthy of running.

- Meisaka